scraper/requirements.txt
pandas
python-dotenv
pymongo